+++ test/test_cursor.py
@@ -366,6 +366,10 @@ class TestCursor(IntegrationTest):
             break
         self.assertRaises(InvalidOperation, a.hint, spec)
 
+        spec = ["num"]
+        first = next(db.test.find().hint(spec))
+        self.assertEqual(99, first.get("num"))
+
     def test_hint_by_name(self):
         db = self.db
         db.test.drop()